+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - scroll the selected account into view in the accounts sidebar #4137
 - dev: clarify scrolling-related code #4121
 - improved performance a bit #4145
+- show contact / group name when pasting invite link in the search field #4151
 
 ### Fixed
 - fix that you can not click header button in dialog when they are on top of the navbar #4093